feat(cli): add mercury cli
