+++ Cargo.toml
@@ -6,6 +6,7 @@ authors = ["Xuejie Xiao <xxuejie@gmail.com>, Eason Gao <kaoimin@qq.com>"]
 edition = "2018"
 
 [dependencies]
+ansi_term = "0.12"
 anyhow = "1.0"
 clap = "2.33"
 # ckb-indexer = { git = "https://github.com/nervosnetwork/ckb-indexer", rev = "c3dbf63" }
@@ -30,8 +31,10 @@ log4rs = "1.0"
 molecule = "=0.6.0"
 num-bigint = "0.4"   
 num-traits = "0.2" 
+regex = "1.5"
 rlp = "0.4"
 rocksdb = { package = "ckb-rocksdb", version = "=0.15.1" }
+rustyline = "8.0"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 tokio = { version = "1.5", features = ["macros", "rt-multi-thread"] }
